@PlayerEG
feat(User): 实现用户借阅图书功能

- 新增 BorrowBookPanel 类，实现用户借阅图书的界面和逻辑
- 更新 User 类，将 book_borrow_id 和 book_borrow_time 字段类型改为 long
- 修改 UserSaveTool 类，同步更新用户借阅相关字段类型
- 在 UserDao 中添加 updateUserBookBorrowInfo 方法，用于更新用户借阅信息
- 优化验证码过期时间判断逻辑

+++ src/main/java/top/playereg/sys/entity/User.java
@@ -15,7 +15,7 @@ public class User {
     private String email;
     private String is_root;
     private String is_del;
-    private int book_borrow_id;
+    private long book_borrow_id;
     private int book_borrow_time;
 
     public User() {
@@ -89,11 +89,11 @@ public class User {
         this.is_del = is_del;
     }
 
-    public int getBook_borrow_id() {
+    public long getBook_borrow_id() {
         return book_borrow_id;
     }
 
-    public void setBook_borrow_id(int book_borrow_id) {
+    public void setBook_borrow_id(long book_borrow_id) {
         this.book_borrow_id = book_borrow_id;
     }
